refactor (lotto): refactor codes(3)
