+++ src/main/kotlin/lotto/Main.kt
@@ -1,13 +1,13 @@
 package lotto
 
+import lotto.model.InputReader
 import lotto.model.LottoMachine
 import lotto.model.Lottos
 import lotto.model.Lotto
-import lotto.model.Money
-import lotto.model.WinningLotto
 import lotto.model.LottoNumber
+import lotto.model.WinningLotto
+import lotto.model.Money
 import lotto.model.Result
-import lotto.view.InputReader
 import lotto.view.InputView
 import lotto.view.ResultView
 import java.math.BigDecimal
@@ -18,9 +18,10 @@ fun main() {
 
     InputView.printBudgetQuestion()
     val budget: Money = inputReader.readBudget()
-    lottoMachine.insertMoney(budget)
+    val availableCount = lottoMachine.insertMoney(budget)
+    ResultView.printLottoCount(availableCount)
+
     val myLottos: Lottos = lottoMachine.buy()
-    ResultView.printLottoCount(lottoMachine)
     ResultView.printMyLottos(myLottos)
 
     InputView.printWinningNumberQuestion()
@@ -28,8 +29,11 @@ fun main() {
 
     InputView.printBonusBallQuestion()
     val bonusNumber: LottoNumber = inputReader.readBonusNumber(winningNumbers)
-    val result: List<Result> = lottoMachine.getResult(WinningLotto(winningNumbers, bonusNumber))
-    val earningRate: BigDecimal = lottoMachine.getEarningRate()
+    val winningLotto = WinningLotto(winningNumbers, bonusNumber)
+
+    val result: List<Result> = lottoMachine.getResult(winningLotto)
     ResultView.printResult(result)
+
+    val earningRate: BigDecimal = lottoMachine.getEarningRate()
     ResultView.printEarningRate(earningRate)
 }
